modules: triggers: Handle all new trigger-on triggers
The 'trigger-on' dictionary contains triggers which could be dictionaries
themselves in case we are using the new style 'trigger-on' triggers.
Each of these dictionaries can either have a boolean or a dictionary
value. For example, the following jjb yaml snippet is translated to the
below dictionaries:

trigger-on:
  - patchset-created-event:
      exclude-drafts: false
      exclude-trivial-rebase: false
      exclude-no-code-change: false
  - change-merged-event: true
  - change-abandoned-event: true
  - change-restored-event:  false
  - draft-published-event:  false

OrderedDict([('patchset-created-event', OrderedDict([('exclude-drafts', False), ('exclude-trivial-rebase', False), ('excluide-no-code-change', False)]))])
OrderedDict([('change-merged-event', True)])
OrderedDict([('change-abandoned-event', True)])
OrderedDict([('change-restored-event', False)])
OrderedDict([('draft-published-event', False)])

The code only handles limited cases of non-six.string_types values such
as 'comment-added-event' and 'patchset-created-event'. The rest of the of
the possible options are silently ignored leading to problematic xml
output:

<triggerOnEvents>
  <com.sonyericsson.hudson.plugins.gerrit.trigger.hudsontrigger.events.PluginPatchsetCreatedEvent>
    <excludeDrafts>false</excludeDrafts>
    <excludeTrivialRebase>false</excludeTrivialRebase>
    <excludeNoCodeChange>false</excludeNoCodeChange>
  </com.sonyericsson.hudson.plugins.gerrit.trigger.hudsontrigger.events.PluginPatchsetCreatedEvent>
</triggerOnEvents>

As we can see, there is no suitable xml output for the
'change-merged-event', 'changed-abandoned-event' etc.

The dictionaries within the 'trigger-on' dictionary only have a single
key so it's easy to handle the rest of the options in a final 'else'
case and append the related element if the option is 'True'. With
the following change, the xml out is the correct one.

<triggerOnEvents>
   <com.sonyericsson.hudson.plugins.gerrit.trigger.hudsontrigger.events.PluginPatchsetCreatedEvent>
     <excludeDrafts>false</excludeDrafts>
     <excludeTrivialRebase>false</excludeTrivialRebase>
     <excludeNoCodeChange>false</excludeNoCodeChange>
   </com.sonyericsson.hudson.plugins.gerrit.trigger.hudsontrigger.events.PluginPatchsetCreatedEvent>
   <com.sonyericsson.hudson.plugins.gerrit.trigger.hudsontrigger.events.PluginChangeMergedEvent/>
   <com.sonyericsson.hudson.plugins.gerrit.trigger.hudsontrigger.events.PluginChangeAbandonedEvent/>
</triggerOnEvents>

This patchset also adds a test case to capture such problems in the future.

Change-Id: I9b1438bb4176a7d35ee47919dbb0cf7a8d2da085