Update patch set 2
Patch Set 2: Code-Review+2

"bug" is probably a bit harsh, it's the way it works to support template usage, it's not something whose behaviour is easily altered to resolve.

Potentially we could use the lazy-loading object always, provided it validates that the referenced file always exists, and then provide something that decides whether to escape the code when performing the templating or not as the case may be.

Certainly non-trivial so this should be sufficient for now.

Patch-set: 2
Reviewer: Gerrit User 1054 <1054@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2