Update patch set 3
Patch Set 3:

(2 comments)

I think this would need examples of what should be getting sorted and also possibly adding of a simple test for the so it's clear how it is intended to sort jobs.

I'm guessing at something along the lines of:

 1-test
 2-abc
 abc
 bcd-1
 bcd

But without throwing some job names at it, I wouldn't be sure.

You're aware that just doing a simple sort the current tests would probably work for the following cases:

 001-name
 002-name
 name

So it's useful to explain how adding the proposed sorting mechanism will sort and is better for users would make it easier for reviewers.

Also should be moved to the utils file.

Patch-set: 3