Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

Could do with some more info in the commit message, I'm also unsure about managing this directly in the library layer and wondering if you meant for this to just affect the output produced by JJB when running it as a tool?

In which case need to look at only changing it at jenkins_jobs/cli/entry.py, however changing it there has the potential to break any tooling calling JJB as a tool, so it may need to be placed behind an option or for an easy override to be provided to users to get the old format.

But I'm kind of speculating as to the intent here, as it's unclear as to which use case it is intending to change.

Patch-set: 1
Label: Code-Review=-1