Update patch set 2
Patch Set 2: Code-Review-1

Should use the convert_mapping_to_xml helper function instead of multiple XML.SubElement calls.

Also I would prefer the minimal and full examples be named "*-minimal.yaml" and "*-full.yaml" so that it's clear which example is the minimal and which one is the full from the filename.

Patch-set: 2
Label: Code-Review=-1