Update patch set 2
Patch Set 2: Workflow-1

Hi, Thanh,

to be honest, I'm not a fan of this administrator priviledges thing either. But in the end I guess this is what most of users do (just for the main user) since jenkins-jobs does not work reliably otherwise (what is more this requirement is not properly documented). So security problem is still there, just with more risk.

I read the whole thread on upstream mailing list you linked and it seems that proper solution is not straightforward and will take a while. In the meantime, what do you about think about making "update" subcommand (delete does not need, does it?) accept "-p" option like currently "test" does:

-p PLUGINS_INFO_PATH  path to plugin info YAML file

While this is not perfect and would require manual work to maintain that YAML file, but at least it would provide an escape route which is under user control. What is more, `jenkins-jobs` could provide "dump-plugins" subcommand which would write that file based on plugin information (which would be useful for "test" as well).

I would be willing to rewrite this patch to add "-p" option if you this is a good idea and maybe follow up with "dump-plugins" later.

Patch-set: 2
Reviewer: Gerrit User 26073 <26073@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Workflow=-1