Add `do-not-fetch-tags` to CloneOption for Git
In one of the recent change to Git plugin behavior[1], tags are
also search as part of branches.

This cause projects that build on all branches to also build on all
tags.

This change allowing for tags to be ignored. The correct XML setting for
this will need to be as follow.

```
    <extensions>
      <hudson.plugins.git.extensions.impl.CloneOption>
        <shallow>false</shallow>
        <noTags>true</noTags>
      </hudson.plugins.git.extensions.impl.CloneOption>
    </extensions>
```

This is the generated result

```
    <extensions>
      <hudson.plugins.git.extensions.impl.CloneOption>
        <shallow>false</shallow>
        <noTags>true</noTags>
      </hudson.plugins.git.extensions.impl.CloneOption>
      <hudson.plugins.git.extensions.impl.WipeWorkspace/>
    </extensions>
```

[1] https://github.com/jenkinsci/git-plugin/pull/340
[2] https://github.com/jenkinsci/git-plugin/commit/bfeda3e661531bd9baec937ec4a4944ac482692c

Change-Id: I2ed2290f9ef8ecd0d9e96aad1d7cbce2964bf2da