Update patch set 2
Patch Set 2:

Hi Modestas,

Sorry for the late reply on this one. Things got really busy this past week.

I didn't realize the update command didn't support the `-p` option. I had thought it did. I would be supportive of this method. As you said we can have a dump of the plugins via a local user account. This file containing the plugin versions can be pushed to the same repo stored with the jjb templates and then reused in a system where we don't allow administrator account access.

Seems like a reasonable workaround while we wait for upstream to give us a proper solution.

Patch-set: 2