Don't lower case bool param defaults
This breaks the case where the default is a parameter that contains
upper case characters.

Why would the default for a parameter be parameterised? In the case
where a block of params is used by multiple jobs, but where they
may need to override the default for some of them.

For example:
  defaults.yml:
    - defaults:
        MY_PARAM: true
  params.yml:
    - parameter:
        name: my_params
        default: "{MY_PARAM}"
    ... other params ...
  job1.yml:
    job:
      parameters:
        - my_params
  job2.yml:
    job:
      parameters:
        - my_params:
            MY_PARAM: false

Change-Id: I76ee956011c81522ee9888ca15d34d33d718ca55