Update patch set 3
Patch Set 3:

I rewrite a lot of this change in order to allow users to configure the logging format. Now is possible to specify it either in the config file or on the command line.

This will allow people to tune the format to meet their needs. 

I really doubt that anyone was relying on the older format that included the timestamp as a prefix. If they would do match lines, they would do it with a regex so is likely to pass the new format which removes the timestamp.

Is ok to change behaviour in v2, so I see no hard reason for keeping the old behaviour since v2 was not released yet.

Patch-set: 3
Label: Verified=0