Update patch set 3
Patch Set 3:

(4 comments)

I think this is close thanks, I have some thoughts on whether it might be better to use a separate namespace for the logging settings, let me know what you think about it.

Unsure whether there is any downside to switching the entire conf reading to RawConfigParser, probably need others to chime on that, one possible workaround is it should be possible to use 'raw=True' for a single value retrieval.

Other thoughts on use of a log_config file are just considering how things might head in the future.

Certainly if we're thinking of changing the default output we should send this out in an RC, as up till now all changes have been internal only with no alterations to external behaviour.

Patch-set: 3
Label: Verified=0