Update patch set 10
Abandoned

Digging into the issue about the condition revealed this functionality is covered by the pipeline publisher added by 6ac3c5087ba6971387cbbd2118bbf4a20fd38f04 which it appears was added so long ago, we completely forgot about it. I'll look to consolidate that one with the trigger_parameterized_builds function so that it has access to the additional settings.

Patch-set: 10
Status: abandoned