Update patch set 4
Patch Set 4:

(4 comments)

The code is mostly fine as is, I've made an alternative suggestion for the main function to use dict, but it's not necessary to switch.

Only items that need to be fixed is clarification as to whether the docstring is missing a default entry for the 'upload' configuration, and also should use the correct exception type.

Patch-set: 4
Reviewer: Gerrit User 1054 <1054@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0