Update patch set 6
Patch Set 6: Code-Review-1

(2 comments)

Spotted two things that should be changed, one this should be added to tests/base.py to replace the existing sort there, so that this gets exercised by all tests, and two should move the sort higher up in the update method and replace the existing sort.

After that good to merge.

Patch-set: 6
Label: Code-Review=-1