Update patch set 1
Patch Set 1:

> LGTM, one question we should change all the plugins default to
 > newest plugin format?

There was some discussions on that on IRC between Darragh and I and I think consensus is yes if work required is low. I've submitted a few patches under this topic and if I'm not incorrect there should only be a few more plugins to convert so I think we should wait until all patches are ready and make the decision then.

Patch-set: 1
Label: Verified=0