(WIP) merge common code
The Run Condition plugin provides common logic for a builder
(Conditional BuildStep) and a publisher (Flexible publish).  Under the
hood the Java implementation is exactly the same but JJB has duplicate
implementation.

Move the logic from the builder to helpers.build_condition, it is the
most complete implementation, notably supports multiple conditions such
as 'And' or 'Or'.
Add a special case for the regex matching which use a slightly different
semantic between plugins:

Builder conditional BuildStep has:

- conditional-step:
    condition-kind: regex-match
    regex: .*
    label: $VAR

Flexible publish has:

- conditional-publisher:
   - condition-kind: regexp
     condition-expression: .*
     condition-searchtext: $VAR

Factor out the doc about condition so it can be included in both the
builder and the publisher: doc/source/helpers/build_condition.rst

Change-Id: Ia78b280963783dcb2e2c21f19173d10e2452a4c2