Add tox to test-requirements.txt.
If I understand correctly, tox is already installed on automated test
nodes; however, the intent here is to improve local development workflow
by ensuring that an appropriate version is installed for running tests
locally. I have installed tox==2.7.0 locally so it seems like a
reasonable version on which to pin the test dependency.

Change-Id: I8052e5792e229b58a48e48bd705ed576acdc2884