Update patch set 17
Patch Set 17:

(2 comments)

If I recall correctly, would this solve the scenario in https://review.openstack.org/473828?

If so we should add some examples as test cases for the improved functionality and include them in the docs.

I do need to look closer at the changed test, wondering if needs to be altered, or if it's just that now these types of failures are caught differently. Guessing it's the second scenario, since we're now expanding the macro and then applying the params as part of an expanded template.

Patch-set: 17
Label: Verified=0