Update patch set 3
Patch Set 3:

@Thanh based on your terminology

Current behaviour:

 1) job-group params:
 a. specific override as defined under the 'jobs' defintion.
 b. job-group top level definition

 2) project params:
 a. specific override as defined under the 'jobs' definition.
 b. project top level definition

 3) default params

 4) job-template params


I guess I can see the simplicity of having project params as a high priority, however at the moment they are one of the lowest priorities so if we move it around need to be sure the change is beneficial in providing flexibility.

However as a first pass do we agree that making this configurable is the correct path to adding it for V2 as opposed to changing the behaviour for existing users?

Patch-set: 3
Label: Verified=0