Update patch set 3
Patch Set 3: Code-Review+2

I can agree with that. If that's how it currently is I guess there's no reason for me to be against that change. I did incorrectly think that project definitions had the highest priority and in my opinion it should because that's the most specific definition imo.

On one of my project's we're using job-groups with parameter definitions to define "default" values and expect projects to override that default value as necessary in their more specific "project" definition if they need to.

Patch-set: 3
Reviewer: Gerrit User 13807 <13807@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2