Update patch set 1
Patch Set 1:

the order of py35 and py27 was to prevent an issue with testr compatibility, as I thought creating the test repository with 2.7 first would result in it being unreadable by testr when run under 3.5, but wasn't an issue the other way around. Is this gone?

Patch-set: 1
Label: Verified=0