Update patch set 8
Patch Set 8:

(1 comment)

I've always hoped we might get around to enhancing JJB by supporting jinja2 as a template engine instead of the current one as it just provides so much more flexibility.

I guess the only missing piece to this one is to confirm that behaviour when using the jinja2 include within a job-template works as expected. I think it does, based on it copying the lazyloader approach, but probably just needs confirmation with another test case and this can help users with some examples.

Patch-set: 8
Label: Verified=0