Update patch set 12
Patch Set 12:

Technically addition of the folder attribute isn't needed, however the rest of the code is needed to ensure both generating the test output creates the correct folder structure locally, as well as retrieving the full job name from the remote jenkins when getting the list of jobs.

Originally we needed to expand the job name as 'job/<folder>/job/<name>' when creating it, so it was convenient to keep it separate.

It might also be handy to be applied from a set of defaults, so depending on your template you may choose a different set of defaults for different teams and some may default the 'folder' attribute and some may not in which case the job template name doesn't need to be aware that it needs a folder name inserted into the 'name' as '{folder}/some-job-type-{name}'

Patch-set: 12
Label: Verified=0