Adds support for prefix-variable in version-number wrapper
prefix-variable is currently not available for use in the version-number wrapper.
updated wrapper for support of this variable, and updated test fixtures to also use this variable.
set default value of prefix-variable to ''
Errors out without a default value set saying the attribute is missing, and since this is an optional attribute we don't want that.

Task: 4885
Change-Id: I151eef271194fb38dfde7518698f2ff06e3a8f01