Add Support of additional-credentials in the SVN-plugin. Fix the comments of codereview
Since the 30.01.2017 the SVN plugin support additional-credentials, but the jenkins-job-builder can't generate this part:

<additionalCredentials>
    <hudson.scm.SubversionSCM_-AdditionalCredentials>
        <realm>...</realm>
        <credentialsId>...</credentialsId>
     </hudson.scm.SubversionSCM_-AdditionalCredentials>
</additionalCredentials>

For the first step it is implemented only for one additional-credentials.
Now you have 2 optional api value additional-credential-id and
additional-credential-realm. Only if both value are set in the *.yml,
it generate the "additionalCredentials" part in the config.xml of the
jenkins-job.

Change-Id: I76460f3d1009680c861f3a72f543facf0cf0d184
Task:4884
Story:2001152