Update patch set 2
Patch Set 2:

Hi Jiri, I understand this can be frustrating and it affects myself as well as it's sometimes it takes unreasonably long to get patches in. Please understand that everyone who works on this project are all volunteers and none of us work on JJB as part of our day jobs so reviews can be slow (unfortunately).

I've found pinging people in #openstack-jjb every once in awhile to remind them to review useful or bumping this patch as you have done. Hopefully someone will see this and review it. I'll try to ping some folks as well and maybe someone will notice.

Patch-set: 2