Add support for unstable return parameter for shell builders
This continues to support the current string form for job configuration,
as well as adding a form that supports a mapping.  This mapping has two
keys: "command", containing the command to be executed (i.e. the current
string parameter); and "unstable-return", containing the exit code that
configures the exit code that should cause a build to be marked
unstable.

Change-Id: I43ecc883236bbf8fc6de7ed8992e6d90da7d48ac