Update patch set 13
Patch Set 13:

From the github project README it was using an early version of this patch along with the change from python-jenkins, but that's about it.

There is another patch around for creating folders that needs some work, with this one and that combined it JJB would support everything natively that was added by the addon.

Patch-set: 13
Reviewer: Gerrit User 1054 <1054@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0