Update patch set 5
Patch Set 5: Code-Review-1

(2 comments)

According to https://wiki.jenkins.io/display/JENKINS/NodeJS+Plugin we should also be able to configure an npmrc file. Should probably update the plugin to support it.

As for tests we are trying to standardize on the test format of creating 2 test files.

nodejs-minimal.yaml
nodejs-full.yaml

one containing the absolute minimun configuration necessary to use the plugin and the other containing all of the possible configurations with a non-default value.

Patch-set: 5
Label: Code-Review=-1