Update patch set 2
Abandoned

I will abandon this path, in short words the problem in the following:
the error occurs when Jenkins_url is instance of unicode and the job content contains non-accii symbols. in this case, when full HTTP request message builds it will be instanse of unicode and the exception occurs when this body is converted to str by using ascii codec.
If it is not issue for you or if you thin the patch is not correct. You are welcome to make correct patch.

Patch-set: 2
Status: abandoned