Update patch set 1
Patch Set 1: Code-Review+1

Is it possibly failing in this case because a specific yaml file was passed in the example which didn't include the metadata along with it?

I ask because the verify jobs in other patches do not fail the basic test case which tells me it's working in master.

Patch-set: 1
Label: Code-Review=+1