Update patch set 1
Patch Set 1:

I think you've cracked the underlying design issue I was concerned about with my initial approach.

Registering the expansion points with a central registry instead of adding to the entrypoints as I was looking at will be much more efficient.

Do we still need to make sure the entrypoints are loaded first though? As otherwise this would only work for the subsequent templates and not for an initial one, as the registration only occurs once the entrypoint triggers the module to be loaded.

Patch-set: 1
Reviewer: Gerrit User 1054 <1054@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0