Update patch set 3
Abandoned

I am abandoning this because of all the spaghetti code convolutions necessary to support an actually quite small number of very dissimilar approaches to inner macro expansion. Instead I have fixed 530419 which removes the MacroRegistry and returns to the previous approach of expanding macros which was to do it during xml expansion using the ModuleRegistry.

Patch-set: 3
Status: abandoned