Update patch set 2
Patch Set 2: Code-Review+1

(1 comment)

Would like to make sure we retain the '_' for the entry_points_cache class attribute, but can be done in a separate patch.

Before we land this, I'd like to get it straight in my head about the benefits that we would have from expanding the macros ahead of conversion.

 * modules become a strict yaml -> xml conversion
 * macros may take data from template defaults instead of needing passthru for all settings

The second one is the more interesting use from my perspective, but I suspect that in addition to the practicalities of getting this fully working, there was possibly additional work in ensuring existing use of macros remained working (bugs rather than design issues).

Do we think it would be possible to restore this within v2 in the future if happy with a subsequent implementation without needing a major version bump?

Patch-set: 2
Label: Code-Review=+1