Update patch set 2
Patch Set 2:

(1 comment)

> (1 comment)
 > 
 > Would like to make sure we retain the '_' for the entry_points_cache
 > class attribute, but can be done in a separate patch.

Okay, I will update this shortly.

 > 
 > Before we land this, I'd like to get it straight in my head about
 > the benefits that we would have from expanding the macros ahead of
 > conversion.
 > 
 > * modules become a strict yaml -> xml conversion

Yeah, this was the ideological driving factor for my original refactor. While attempting to make the decorator-based approach applicable in all inner macro expansion cases I realize that the code I was writing was ultimately not very approachable for others who may work on this in the future and that it would be best to discard it for now.

 > * macros may take data from template defaults instead of needing
 > passthru for all settings

Ah yeah, I forgot about this. This may be worth the trouble, but I think a 2.0 release is more pressing right now.

 > 
 > The second one is the more interesting use from my perspective, but
 > I suspect that in addition to the practicalities of getting this
 > fully working, there was possibly additional work in ensuring
 > existing use of macros remained working (bugs rather than design
 > issues).

Yeah in the test cases I've written as a follow-up to this (https://review.openstack.org/#/c/530598/1) I discovered that the variations between use cases of inner macro expansion required the addition of code paths that even I found difficult to reason about at the time.

 > 
 > Do we think it would be possible to restore this within v2 in the
 > future if happy with a subsequent implementation without needing a
 > major version bump?

It would definitely be possible; we would just have to revert this revert then revive https://review.openstack.org/#/c/530530 and its dependent commit. This is one of the reasons I made addition of test cases a separate change--so someone could revert this without removing those test cases.

Come to think of it, I still a considerable amount of work in my local copy of the repo for 530530. I'll look into updating that soon.

Patch-set: 2
Label: Workflow=0