Update patch set 1
Patch Set 1:

Hi Thanh,

I did not investigate this issue based on the test suite (verify jobs) but rather on the outcome on a clean fresh installation (venv).

You might be in a better position to inspect why the jobs are not failing or why on a base install the invocation fails as described in the commit message.

Thanks

Patch-set: 1