Update patch set 1
Patch Set 1:

(4 comments)

Don't really have enough about what is meant about normal test execution to see the difference this patch makes, could you give an example in the review comments of the difference.

Is it common to make that sort of clean up change in the project code as normally I would only want to do that after a checkout.

Patch-set: 1