Update patch set 2
Patch Set 2:

Hi Jonathan,

 > I would argue that the disparity between the doc and actual
 > behaviour still need to be fixed somewhat. Commit message would
 > obviously need to be updated.
 > 
 > Thoughts?

The [doc] also explicitly says that the default is True, so I don't think you fixed the disparity, you just moved it one sentence further. :-)

I'm not sure what the "If no configuration files are found" sentence is supposed to mean, but it doesn't appear to be true either. Perhaps it refers to the [plugins_info.yaml] "configuration" file that is passed to update -p? If someone knew how it behaves exactly and rewrote this sentence (or the entire doc for query_plugins_info), that would be awesome!

Anyway, the doc is now just as incorrect as it was. If you think that it makes sense to keep the default to False, the doc should be updated to say that the default is False, and I found two tests that now don't make sense any more:
https://git.openstack.org/cgit/openstack-infra/jenkins-job-builder/tree/tools/configs/non-defaults.conf?h=refs/heads/master&id=30d61a21e81e274ba40c33c8279f8062f9eeb373#n13
https://git.openstack.org/cgit/openstack-infra/jenkins-job-builder/tree/tests/cmd/fixtures/disable-query-plugins.conf?h=refs/heads/master&id=30d61a21e81e274ba40c33c8279f8062f9eeb373#n3

[doc]: https://git.openstack.org/cgit/openstack-infra/jenkins-job-builder/tree/doc/source/execution.rst?h=refs/heads/master&id=30d61a21e81e274ba40c33c8279f8062f9eeb373#n78
[plugins_info.yaml]: https://docs.openstack.org/infra/jenkins-job-builder/quick-start.html#use-case-5-providing-plugins-info

Patch-set: 2