Update patch set 3
Patch Set 3:

(6 comments)

Thanks, I tested and have a suggestion for ensuring that by default all tests provide a default empty config file in place of fixing the current single class as otherwise there might be other config settings that could affect other test behaviour for other test scenarios that currently don't mock the same config look up.

Usage of the PYTHONDONTWRITEBYTECODE looks handy, and also side steps a concern I had with 'find' not being present on other systems (or just generally needing to rely on non python programs being present).

Are you thinking that we can release with 'find' for a while to clean up dev envs and then remove as the tests will no longer generate .pyc files?

Last step is to try and capture how changing OS_STDERR_CAPTURE/OS_STDOUT_CAPTURE affects the failed test output as I've not been able to see a difference even when forcing a test to fail that would result in a traceback. They always appear to be captured and displayed with the failed test, so I'm not sure what the expected change in behaviour is supposed to be?

Patch-set: 3
Label: Verified=0