bugfix: skip_vote will always be 'true'
If we define a trigger as a macro, and pass a parameter to define whether
skip-vote or not, the value of the skip-vote is always 'true'. It is used as
str to expand the macro, so we need to change the parameter to boolean first.

example:

- trigger:
    name: 'test'
    triggers:
      - gerrit:
          skip-vote:
            successful: '{is-skip}'

- job:
    name: 'test'

    triggers:
      - 'test':
          is-skip: false

When the job expand the trigger marco, the skip-vote: successful is always 'true',
not the value 'false' which we passed in the job definition.

Change-Id: Ic2438c09c5db90f6c3d041b310ec47027f0bbe60
Signed-off-by: dongwenjuan <dong.wenjuan@zte.com.cn>