Update patch set 1
Patch Set 1:

In macros if you want YAML to retain the datatype you need to pass it as {obj:is-skip}. Otherwise everything will be passed in as a string.

This get's more complicated when more complex data types like dictionaries and lists are passed in so {obj:var} is the best way to handle that imo.

Patch-set: 1