Update patch set 5
Patch Set 5:

(3 comments)

Bit unsure about the use of the mock to resolve the conf file issue instead of returning an empty conf as the else path in get_scenarios. Will need to review the other code to be sure it doesn't have potential for unexpected side effects on the code that tests the config object.

Patch-set: 5
Label: Verified=0