Update patch set 1
Patch Set 1:

you can use 'obj:is-skip' to pass through the original object, otherwise we would need to change all true/false value readings everywhere in the codebase.

Might be worth discussing in the irc chat room as it might be worth considering doing this everywhere in the code if we are going to do it anywhere.

Patch-set: 1