Update patch set 5
Patch Set 5:

This patch is the result of discussion ongoing here:

https://review.openstack.org/#/c/527804/

Note that change 527804 changed the default behaviour for all use cases where this patch propose to revert to old behaviour and only change the querying to false in the special case where no conf file is found. This actually fit the current documentation.

Patch-set: 5