Update patch set 6
Patch Set 6:

Had a chat about this in the IRC channel to consider the impact and also whether there was a way to achieve the desired outcome with a lower impact. Obviously we don't want to break existing users, and the switch to not requiring querying plugins was due to Jenkins V2.

Suggestion is to look to add support for 3 states for querying plugins where query_plugins_info =>

 False: skip querying for plugins
 None: query for plugins but don't fail on authentication error
 True: query for plugins and fail if don't succeed

Making 'None' the default that can be overridden by user configs should mean that the default changing with the next release won't break anyone working with V2 already and at the same time means that it at least attempts to query the plugins by default, but it can be disabled or made a hard requirement by users as desired.

Does that make sense?

Patch-set: 6
Reviewer: Gerrit User 1054 <1054@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0