Add support for plugin column views
Add support plugin column views. The default columns are
added which dont require additional plugin installed.

For columns which require plugin support they are
added only if the column is provided as in input.

The following plugins are supported with view columns:

- Robot
- Findbugs
- JaCoCo
- Git
- ScheduleBuild
- BuildFilter

Also added a column for the description which does not require
additional plugin support.

Change-Id: I5d756e5741b0aa1ca58bcc69b79e8e3436e9759d
Signed-off-by: Anil Belur <askb23@gmail.com>