Update patch set 1
Patch Set 1: Code-Review-1

Can we make this plugin support the old format as well for folks who have not and / or cannot update to the latest and greatest?

Take a look at the slack publisher plugin for an example on how to support multiple versions of a plugin.

Patch-set: 1
Reviewer: Gerrit User 13807 <13807@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1