Update patch set 1
Patch Set 1:

> I have updated the branch but unfortunately following the
 > instructions [1] it has created a new PR [2].

Ah that's because you changed the "Change-Id" field. That is the line that allows Gerrit to know if it needs to update an existing patch or not. If you'd like to update this patch and abandon the other one then set the change-id field in the git commit message to the one from this patch.

Patch-set: 1