Update patch set 8
Patch Set 8: Code-Review-1

If you are reading the local *.yaml files to list all the jobs 
why is the `-p` option required? I tried using the list option and does not work without passing `--conf`. I think its not required to mandate a connection to the server when you are parsing the local yaml files and listing the jobs.

Patch-set: 8