Update patch set 14
Patch Set 14:

main reason why this does not use PYTHONHTTPSVERIFY is that when I opened this issue ~3 years ago that variable did not exist.

If that works now in most distributions and okey to be a global switch rather than just in regard to the jenkins connection then that will suffice.

Patch-set: 14