Update patch set 24
Patch Set 24:

(11 comments)

So it seems that the scm part is only partially common, and that is mostly just the advanced options for repository types that are git based.

matching the style of the code in the rest of the repository would be a good idea, I found it jarring reviewing as I was expecting the code to flow in a particular manner.

The test code should also be tweaked to follow the same approach as the other test classes, guessing it was just the need to mock the uuid response that might have caused the use of the cli test command entry point, but mocks can be applied on a class level using decorators as well.

Patch-set: 24
Label: Verified=0