Update patch set 26
Patch Set 26:

(7 comments)

I've put up a suggested example using GitHub of how the test code can be altered to make use of the existing classes to ensure a better output in the case of differences.

Some more questions & suggested tidy up of the code.

A good start on the generic blacklisting of modules, in the short term it might be worth just going with the original hardcoded approach to skipping the SCM module for the multibranch so we're not trying to trash out the best approach and blocking this change from landing. So if that was in a patch layered on top of this one, we can land this once the few minor items are fixed up.

Patch-set: 26