Update patch set 14
Patch Set 14:

(1 comment)

It might still be worth supporting within python-jenkins on a per connection basis, as the default context here as well as the env variable only work on a global basis. This means that any scripts that happen to contact multiple instances using jjb as an API can only disable it for all of them or none unless using something like multiprocess to execute jjb.

I'd also note that this only appeared in python 2.7.12, so likely most RHEL/CentOS 7.x installations do not have a new enough python.

http://mirror.centos.org/centos/7.4.1708/os/x86_64/Packages/python-2.7.5-58.el7.x86_64.rpm -> note I didn't check to see if the brought the patch back as sometimes RedHat does that.

Any interest on building on the requests patch @ https://review.openstack.org/205637 to be able to enable/disable SSL verification for those not yet in a position to move to a new enough python?

Patch-set: 14