Update patch set 34
Patch Set 34: Code-Review+1

Alright. I reworked this entire patch from top to bottom to make it conform to our current code style guidelines and reviewed each SCM component to make sure every parameter being provided was actually configuring something in the UI. There were many parameters that were removed as they did not appear to change anything which implies that they are obsolete config settings.

I broke off the Gerrit SCM to a different patch Ib0c3a6d179e11b91c7a34f3385fd947413291785 as I realized it is not a released plugin and would prefer if we did not land things in JJB that do not have an official release.

Docs were moved around to make things easier to read as well and the different SCM modules were migrated to *_scm() functions out of the code.

Please review this patch with fresh eyes as it was modified in a fairly major way.

Patch-set: 34
Label: Code-Review=+1