Update patch set 38
Patch Set 38:

(9 comments)

Couple of questions, looks like the tests answers one of them for me. Only nits spotted were the docstring for a couple of options needs to be updated to 'optional' to match the code and one spot where we might want to raise an exception that reports to the user the correct values allowed.

Patch-set: 38
Label: Verified=0