Update patch set 2
Patch Set 2:

> @Thanh Ha Yeah that's fine by me. I'm very new to this code base so
 > I'm very open to any better way this could be improved. I'll try
 > that out tomorrow... Thanks for the quick review!

No worries, we are in a kind of in between as we have not migrated all of the code to the new syntax / style yet so there's a mix of old and new styles.

Look for "mapping" variables in other plugins in the modules directory. There's plenty of examples. The mapping is a list of tuples so you can pass many parameters to it and pass it to convert_mapping_to_xml to handle it. The tuple looks something like:

 
 (yaml_input, xml_output, default_value)
 

Optionally for parameters that have a list of valid values you can also pass a 4th parameter which is a list of the valid values for that setting.

Patch-set: 2