Update patch set 3
Patch Set 3:

> Thank you Thanh Ha! Would you mind tagging and pushing a new
 > release with this fix? Users were asking for it in
 > https://github.com/ktdreyer/jenkins-job-wrecker/pull/73

Absolutely. I just pushed out a 2.0.5 so should appear in pypi shortly.

For future reference this patch missed the 2.0.4 release that went out yesterday but we typically release every 2 weeks give or take a week. Since January we've been trying to get release momentum up in this project.

If you have patches in the future that you'd like to make it in a release window please join us on IRC in #openstack-jjb @ irc.freenode.net and let us know which patches you'd like priority on.

Cheers,
Thanh

Patch-set: 3